cat-service-diagnostic/requirements.txt
airefinery-sdk
fastapi
uvicorn
python-dotenv
pandas
